pyproject.toml
[project]
name = "hat-format-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]

[tool.mypy]
python_version = "3.12"
files = ["app/**/*.py"]
plugins = ["pydantic.mypy"]
strict = true

[tool.ruff]
target-version = "py312"
include = ["app/**/*.py"]

[tool.uv]
cache-dir=".cache"
